Some fixes, Some changes.
I have now swapped to resources for providing player data! These will be easier to manage, And i want to keep them working with the battle members, Which are gonna be inheriting the script for turn based combat.

I have also split the enemy and player on the gui! So you can see both their health and name.